+++ gxexternalproviders/src/main/java/com/genexus/db/driver/ExternalProviderService.java
@@ -11,6 +11,11 @@ public abstract class ExternalProviderService {
 
 	abstract String getName();
 
+	static final String DEFAULT_ACL = "DEFAULT_ACL";
+	static final String DEFAULT_EXPIRATION = "DEFAULT_EXPIRATION";
+	static final String FOLDER = "FOLDER_NAME";
+
+	static final int DEFAULT_EXPIRATION_MINUTES = 24 * 60;
 	public ExternalProviderService() {
 
 	}
@@ -22,7 +27,7 @@ public abstract class ExternalProviderService {
 	public String getEncryptedPropertyValue(String propertyName, String alternativePropertyName) throws Exception {
 		String value = getEncryptedPropertyValue(propertyName, alternativePropertyName, null);
 		if (value == null) {
-			String errorMessage = String.format("Service configuration error - Property name %s must be defined", propertyName);
+			String errorMessage = String.format("Service configuration error - Property name %s must be defined", resolvePropertyName(propertyName));
 			logger.fatal(errorMessage);
 			throw new Exception(errorMessage);
 		}
@@ -36,7 +41,7 @@ public abstract class ExternalProviderService {
 				value = Encryption.decrypt64(value);
 			}
 			catch (Exception e) {
-				logger.warn("Could not decrypt property name: " + propertyName);
+				logger.warn("Could not decrypt property name: " + resolvePropertyName(propertyName));
 			}
 		}
 		return value;
@@ -45,7 +50,7 @@ public abstract class ExternalProviderService {
 	public String getPropertyValue(String propertyName, String alternativePropertyName) throws Exception{
 		String value = getPropertyValue(propertyName, alternativePropertyName, null);
 		if (value == null) {
-			String errorMessage = String.format("Service configuration error - Property name %s must be defined", propertyName);
+			String errorMessage = String.format("Service configuration error - Property name %s must be defined", resolvePropertyName(propertyName));
 			logger.fatal(errorMessage);
 			throw new Exception(errorMessage);
 		}
@@ -53,6 +58,7 @@ public abstract class ExternalProviderService {
 	}
 
 	public String getPropertyValue(String propertyName, String alternativePropertyName, String defaultValue) {
+		propertyName = resolvePropertyName(propertyName);
 		String value = System.getenv(propertyName);
 		if (value == null || value.length() == 0){
 			value = System.getenv(alternativePropertyName);
@@ -64,5 +70,9 @@ public abstract class ExternalProviderService {
 			}
 		}
 		return value != null? value: defaultValue;
+	}
+
+	private String resolvePropertyName(String propertyName) {
+		return String.format("STORAGE_%s_%s", getName(), propertyName);
 	}
 }
